version 2.0.0
Change all of the thruk_local.conf settings to hashes, so that it is more configurable.
Upgrade thruk to 2.00-2.
Force apache restart before thruk service starts, so that recipes converge on first run.
Add default file so that init script works with SSL redirect.
Fix for fcgid module for Apache 2.4 on CentOS.
Directory permissions for Apache 2.4.
Allow loading of SSL certificates and icons from different cookbooks.
Integration tests for thruk with SSL.
Integration tests and fixes for CentOS 7.1, Debian 7.8, Debian 8.1, Ubuntu 12.04 and Ubuntu 15.04.